% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_annotate_empty.R
\name{mfs_annotate_empty}
\alias{mfs_annotate_empty}
\title{Place an annotation on the least busy corner of a plot}
\usage{
mfs_annotate_empty(s, xs, ys)
}
\arguments{
\item{s}{A text string for use as an annotation}

\item{xs}{The list of x values used in the plot}

\item{ys}{The list of y values used in the plot}
}
\description{
This function returns a ggplot2 annotation grob for addition to an existing
    ggplot object. The annotation should have coordinates and text
    justification appropriate for placing it in the least crowded corner
    of the existing plot, based on x and y lists provided.
}
\examples{
ggplot2(...) + mfs_annotate_empty("a comment", data[,"y"], data[,"x"])
}
\keyword{annotate}
\keyword{ggplot2}
\keyword{text}
