% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_color_scheme.R
\name{mfs_color_scheme}
\alias{mfs_color_scheme}
\title{A function that returns vectors of pre-selected color gradients}
\usage{
mfs_color_scheme(color_scheme = "blackish")
}
\arguments{
\item{color_scheme}{Optionally, a color scheme for the plot, "blackish" "bluish" "reddish" "greenish"}
}
\description{
This function simply returns canned vector of colors
}
\examples{
my_colors = mfs_color_scheme("bluish")
}
\keyword{colors}
