% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_glm_string.R
\name{mfs_glm_string}
\alias{mfs_glm_string}
\title{Return a concise string description of a linear model}
\usage{
mfs_glm_string(model, s.term = "", s.units = "", s.form = "\%0.2f",
  p.level = 0.05, do.p = TRUE)
}
\arguments{
\item{model}{A generalized linear model, returned from glm()}

\item{s.term}{Optionally, specify which variable in your model you are reporting. It must match the model contents.}

\item{s.units}{Optionally, report units as a string, like "mm" or "lbs/in"}

\item{s.form}{a sprintf string, default "\%0.2f", to indicate the format of the numeric output}

\item{p.level}{If you require a p-threshold other than 0.05, set it here}

\item{do.p}{If you don't want p-value reported, set this to FALSE}
}
\description{
This function queries a generalized linear model and produces a string
    to succinctly quantify the results. Useful for annotating plots.
}
\examples{
mfs_glm_string(glm(y~x,data=z), p.level=0.01)
}
\keyword{confidence}
\keyword{interval}
\keyword{p-value}
\keyword{string}
\keyword{summary}
