% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_histo_density.R
\name{mfs_histo_density}
\alias{mfs_histo_density}
\title{A Layered Histogram Function}
\usage{
mfs_histo_density(d, whichcol, xlab = "__", title = "__",
  color_scheme = "blackish")
}
\arguments{
\item{d}{A data.frame containing the column of interest}

\item{whichcol}{The name of the column of interest}

\item{xlab}{Optionally, provide a label for the x axis}

\item{title}{Optionally, provide a title to override the default}

\item{color_scheme}{Optionally, a color scheme for the plot, "blackish" "bluish" "reddish" "greenish"}
}
\description{
This function uses ggplot2 to create a histogram with a background normal
    curve and a foreground density curve.
}
\examples{
mfs_histo_density(d=my_dataframe, whichcol="colname")
mfs_histo_density(my_dataframe, "colname", title="Look, it's normal!")
}
\keyword{density}
\keyword{distribution}
\keyword{histogram}
\keyword{normality}
