% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_scatter_from_models.R
\name{mfs_scatter_from_models}
\alias{mfs_scatter_from_models}
\title{A Lined Scatterplot Function}
\usage{
mfs_scatter_from_models(model1, model2, xlab = "", y1lab = "", y2lab = "",
  title = "")
}
\arguments{
\item{model1}{A linear model, returned from glm()}

\item{model2}{A linear model, returned from glm()}

\item{xlab}{Optionally, provide a label for the x axis}

\item{y1lab}{Optionally, provide a label for the y axis}

\item{y2lab}{Optionally, provide a label for the y axis}

\item{title}{Optionally, provide a title to override the default}
}
\description{
This function uses ggplot2 to create a multi-scatterplot with a line of best fit
    and a loess smoother overlaid, based on two existing linear models. It is
    often discouraged in strong terms to do this, so this library requires that
    the x axes in both models are identical before proceeding. With identical x
    axes, I don't see the problem in viewing two dependent variables together.
}
\examples{
mfs_scatter_from_models(glm(y~x,data=d), glm(z~x,data=d))
mfs_scatter_from_models(glm(y~x,data=d), glm(z~x,data=d), title="A great plot")
mfs_scatter_from_models(glm(y~x,data=d), glm(z~x,data=d), xlab="Independent variable")
}
\keyword{glm}
\keyword{loess}
\keyword{plot}
\keyword{scatter}
