% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs_scatter_from_model.R
\name{mfs_scatter_from_model}
\alias{mfs_scatter_from_model}
\title{A Lined Scatterplot Function}
\usage{
mfs_scatter_from_model(model, xlab = "", ylab = "", title = "",
  color_scheme = "blackish")
}
\arguments{
\item{model}{A linear model, returned from glm()}

\item{xlab}{Optionally, provide a label for the x axis}

\item{ylab}{Optionally, provide a label for the y axis}

\item{title}{Optionally, provide a title to override the default}

\item{color_scheme}{Optionally, a color scheme for the plot, "blackish" "bluish" "reddish" "greenish"}
}
\description{
This function uses ggplot2 to create a scatterplot with a line of best fit
    and a loess smoother overlaid, based on an existing linear model.
}
\examples{
mfs_scatter_from_model(glm(y~x,data=z))
mfs_scatter_from_model(glm(y~x,data=z), title="A great plot")
mfs_scatter_from_model(glm(y~x,data=z), xlab="Independent variable")
}
\keyword{glm}
\keyword{loess}
\keyword{plot}
\keyword{scatter}
